+++ src/from.rs
@@ -13,6 +13,7 @@
 //         Self::try_from(t).unwrap()
 //     }
 // }
+// See <https://github.com/rust-lang/rust/issues/50133>
 
 // FEATURE: (BLOCKED) It would be nice if we could make TryFrom assignment work
 // for all Uints.
@@ -31,85 +32,306 @@
 // }
 
 use crate::Uint;
-use core::{any::type_name, convert::TryFrom, fmt::Display, marker::PhantomData};
+use core::{any::type_name, convert::TryFrom, fmt::Debug};
 use thiserror::Error;
 
+/// Error for [`TryFrom<T>`][TryFrom] for [`Uint`].
 #[derive(Clone, Copy, Debug, Error, Eq, PartialEq, Hash)]
-pub enum ToUintError {
+pub enum ToUintError<T> {
+    /// Value is too large to fit the Uint.
+    ///
+    /// `.0` is `BITS` and `.1` is the wrapped value.
     #[error("Value is too large for Uint<{0}>")]
-    ValueTooLarge(usize),
+    ValueTooLarge(usize, T),
 
+    /// Negative values can not be represented as Uint.
+    ///
+    /// `.0` is `BITS` and `.1` is the wrapped value.
     #[error("Negative values can not be represented as Uint<{0}>")]
-    ValueNegative(usize),
+    ValueNegative(usize, T),
 
+    /// 'Not a number' (NaN) not be represented as Uint
     #[error("'Not a number' (NaN) not be represented as Uint<{0}>")]
     NotANumber(usize),
 }
 
+/// Error for [`TryFrom<Uint>`][TryFrom].
 #[allow(clippy::derive_partial_eq_without_eq)] // False positive
 #[derive(Clone, Copy, Debug, Error, PartialEq, Eq, Hash)]
 #[allow(clippy::module_name_repetitions)]
-pub enum FromUintError<const BITS: usize, T> {
-    #[error("Uint<{}> value is too large for {}", BITS, type_name::<T>())]
-    Overflow(PhantomData<T>),
+pub enum FromUintError<T> {
+    /// The Uint value is too large for the target type.
+    ///
+    /// `.0` number of `BITS` in the Uint, `.1` is the wrapped value and
+    /// `.2` is the maximum representable value in the target type.
+    #[error("Uint<{0}> value is too large for {}", type_name::<T>())]
+    Overflow(usize, T, T),
 }
 
+/// Error for [`TryFrom<Uint>`][TryFrom] for [`ark_ff`](https://docs.rs/ark-ff) and others.
 #[allow(dead_code)] // This is used by some support features.
 #[derive(Debug, Clone, Copy, Error)]
 pub enum ToFieldError {
+    /// Number is equal or larger than the target field modulus.
     #[error("Number is equal or larger than the target field modulus.")]
     NotInField,
 }
 
 impl<const BITS: usize, const LIMBS: usize> Uint<BITS, LIMBS> {
+    /// Construct a new [`Uint`] from the value.
+    ///
     /// # Panics
     ///
     /// Panics if the conversion fails, for example if the value is too large
     /// for the bit-size of the [`Uint`]. The panic will be attributed to the
     /// call site.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use ruint::{Uint, uint, aliases::*};
+    /// # uint!{
+    /// assert_eq!(U8::from(142_u16), 142_U8);
+    /// assert_eq!(U64::from(0x7014b4c2d1f2_U256), 0x7014b4c2d1f2_U64);
+    /// assert_eq!(U64::from(3.145), 3_U64);
+    /// # }
+    /// ```
     #[must_use]
     #[track_caller]
     pub fn from<T>(value: T) -> Self
     where
-        Self: TryFrom<T>,
-        <Self as TryFrom<T>>::Error: Display,
+        Self: UintTryFrom<T>,
     {
-        match Self::try_from(value) {
-            Ok(uint) => uint,
+        match Self::uint_try_from(value) {
+            Ok(n) => n,
             Err(e) => panic!("Uint conversion error: {}", e),
         }
     }
 
+    /// Construct a new [`Uint`] from the value saturating the value to the
+    /// minimum or maximum value of the [`Uint`].
+    ///
+    /// If the value is not a number (like `f64::NAN`), then the result is
+    /// set zero.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use ruint::{Uint, uint, aliases::*};
+    /// # uint!{
+    /// assert_eq!(U8::saturating_from(300_u16), 255_U8);
+    /// assert_eq!(U8::saturating_from(-10_i16),   0_U8);
+    /// assert_eq!(U32::saturating_from(0x7014b4c2d1f2_U256), U32::MAX);
+    /// # }
+    /// ```
     #[must_use]
-    pub fn checked_from_uint<const BITS_SRC: usize, const LIMBS_SRC: usize>(
-        value: Uint<BITS_SRC, LIMBS_SRC>,
-    ) -> Option<Self> {
-        Self::checked_from_limbs_slice(value.as_limbs())
+    pub fn saturating_from<T>(value: T) -> Self
+    where
+        Self: UintTryFrom<T>,
+    {
+        match Self::uint_try_from(value) {
+            Ok(n) => n,
+            Err(ToUintError::ValueTooLarge(..)) => Self::MAX,
+            Err(ToUintError::ValueNegative(..) | ToUintError::NotANumber(_)) => Self::ZERO,
+        }
     }
 
+    /// Construct a new [`Uint`] from the value saturating the value to the
+    /// minimum or maximum value of the [`Uint`].
+    ///
+    /// If the value is not a number (like `f64::NAN`), then the result is
+    /// set zero.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use ruint::{Uint, uint, aliases::*};
+    /// # uint!{
+    /// assert_eq!(U8::wrapping_from(300_u16),  44_U8);
+    /// assert_eq!(U8::wrapping_from(-10_i16), 246_U8);
+    /// assert_eq!(U32::wrapping_from(0x7014b4c2d1f2_U256), 0xb4c2d1f2_U32);
+    /// # }
+    /// ```
+    #[must_use]
+    pub fn wrapping_from<T>(value: T) -> Self
+    where
+        Self: UintTryFrom<T>,
+    {
+        match Self::uint_try_from(value) {
+            Ok(n) | Err(ToUintError::ValueTooLarge(_, n) | ToUintError::ValueNegative(_, n)) => n,
+            Err(ToUintError::NotANumber(_)) => Self::ZERO,
+        }
+    }
+
+    /// # Panics
+    ///
+    /// Panics if the conversion fails, for example if the value is too large
+    /// for the bit-size of the target type.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use ruint::{Uint, uint, aliases::*};
+    /// # uint!{
+    /// assert_eq!(300_U12.to::<i16>(), 300_i16);
+    /// assert_eq!(300_U12.to::<U256>(), 300_U256);
+    /// # }
+    /// ```
+    #[must_use]
+    #[track_caller]
+    pub fn to<T>(&self) -> T
+    where
+        Self: UintTryTo<T>,
+        T: Debug,
+    {
+        self.uint_try_to().expect("Uint conversion error")
+    }
+
+    /// # Examples
+    ///
+    /// ```
+    /// # use ruint::{Uint, uint, aliases::*};
+    /// # uint!{
+    /// assert_eq!(300_U12.wrapping_to::<i8>(), 44_i8);
+    /// assert_eq!(255_U32.wrapping_to::<i8>(), -1_i8);
+    /// assert_eq!(0x1337cafec0d3_U256.wrapping_to::<U32>(), 0xcafec0d3_U32);
+    /// # }
+    /// ```
+    #[must_use]
+    pub fn wrapping_to<T>(&self) -> T
+    where
+        Self: UintTryTo<T>,
+    {
+        match self.uint_try_to() {
+            Ok(n) | Err(FromUintError::Overflow(_, n, _)) => n,
+        }
+    }
+
+    /// # Examples
+    ///
+    /// ```
+    /// # use ruint::{Uint, uint, aliases::*};
+    /// # uint!{
+    /// assert_eq!(300_U12.saturating_to::<i16>(), 300_i16);
+    /// assert_eq!(255_U32.saturating_to::<i8>(), 127);
+    /// assert_eq!(0x1337cafec0d3_U256.saturating_to::<U32>(), U32::MAX);
+    /// # }
+    /// ```
+    #[must_use]
+    pub fn saturating_to<T>(&self) -> T
+    where
+        Self: UintTryTo<T>,
+    {
+        match self.uint_try_to() {
+            Ok(n) | Err(FromUintError::Overflow(_, _, n)) => n,
+        }
+    }
+
+    /// Construct a new [`Uint`] from a potentially different sized [`Uint`].
+    ///
     /// # Panics
     ///
     /// Panics if the value is too large for the target type.
     #[must_use]
     #[track_caller]
+    #[deprecated(since = "1.4.0", note = "Use `::from()` instead.")]
     pub fn from_uint<const BITS_SRC: usize, const LIMBS_SRC: usize>(
         value: Uint<BITS_SRC, LIMBS_SRC>,
     ) -> Self {
         Self::from_limbs_slice(value.as_limbs())
     }
+
+    #[must_use]
+    #[deprecated(since = "1.4.0", note = "Use `::checked_from()` instead.")]
+    pub fn checked_from_uint<const BITS_SRC: usize, const LIMBS_SRC: usize>(
+        value: Uint<BITS_SRC, LIMBS_SRC>,
+    ) -> Option<Self> {
+        Self::checked_from_limbs_slice(value.as_limbs())
+    }
+}
+
+/// ⚠️ Workaround for [Rust issue #50133](https://github.com/rust-lang/rust/issues/50133).
+/// Use [`TryFrom`] instead.
+///
+/// We cannot implement [`TryFrom<Uint>`] for [`Uint`] directly, but we can
+/// create a new identical trait and implement it there. We can even give this
+/// trait a blanket implementation inheriting all [`TryFrom<_>`]
+/// implementations.
+#[allow(clippy::module_name_repetitions)]
+pub trait UintTryFrom<T>: Sized {
+    #[allow(clippy::missing_errors_doc)]
+    fn uint_try_from(value: T) -> Result<Self, ToUintError<Self>>;
+}
+
+/// Blanket implementation for any type that implements [`TryFrom<Uint>`].
+impl<const BITS: usize, const LIMBS: usize, T> UintTryFrom<T> for Uint<BITS, LIMBS>
+where
+    Self: TryFrom<T, Error = ToUintError<Self>>,
+{
+    fn uint_try_from(value: T) -> Result<Self, ToUintError<Self>> {
+        Self::try_from(value)
+    }
+}
+
+impl<const BITS: usize, const LIMBS: usize, const BITS_SRC: usize, const LIMBS_SRC: usize>
+    UintTryFrom<Uint<BITS_SRC, LIMBS_SRC>> for Uint<BITS, LIMBS>
+{
+    fn uint_try_from(value: Uint<BITS_SRC, LIMBS_SRC>) -> Result<Self, ToUintError<Self>> {
+        let (n, overflow) = Self::overflowing_from_limbs_slice(value.as_limbs());
+        if overflow {
+            Err(ToUintError::ValueTooLarge(BITS, n))
+        } else {
+            Ok(n)
+        }
+    }
+}
+
+/// ⚠️ Workaround for [Rust issue #50133](https://github.com/rust-lang/rust/issues/50133).
+/// Use [`TryFrom`] instead.
+pub trait UintTryTo<T>: Sized {
+    #[allow(clippy::missing_errors_doc)]
+    fn uint_try_to(&self) -> Result<T, FromUintError<T>>;
+}
+
+impl<const BITS: usize, const LIMBS: usize, T> UintTryTo<T> for Uint<BITS, LIMBS>
+where
+    T: for<'a> TryFrom<&'a Self, Error = FromUintError<T>>,
+{
+    fn uint_try_to(&self) -> Result<T, FromUintError<T>> {
+        T::try_from(self)
+    }
+}
+
+impl<const BITS: usize, const LIMBS: usize, const BITS_DST: usize, const LIMBS_DST: usize>
+    UintTryTo<Uint<BITS_DST, LIMBS_DST>> for Uint<BITS, LIMBS>
+{
+    fn uint_try_to(
+        &self,
+    ) -> Result<Uint<BITS_DST, LIMBS_DST>, FromUintError<Uint<BITS_DST, LIMBS_DST>>> {
+        let (n, overflow) = Uint::overflowing_from_limbs_slice(self.as_limbs());
+        if overflow {
+            Err(FromUintError::Overflow(BITS_DST, n, Uint::MAX))
+        } else {
+            Ok(n)
+        }
+    }
 }
 
 // u64 is a single limb, so this is the base case
 impl<const BITS: usize, const LIMBS: usize> TryFrom<u64> for Uint<BITS, LIMBS> {
-    type Error = ToUintError;
+    type Error = ToUintError<Self>;
 
     fn try_from(value: u64) -> Result<Self, Self::Error> {
-        if Self::LIMBS <= 1 {
+        if LIMBS <= 1 {
             if value > Self::MASK {
-                return Err(ToUintError::ValueTooLarge(BITS));
+                let mut limbs = [0; LIMBS];
+                if LIMBS == 1 {
+                    limbs[0] = value & Self::MASK;
+                }
+                return Err(ToUintError::ValueTooLarge(BITS, Self::from_limbs(limbs)));
             }
-            if Self::LIMBS == 0 {
-                return Ok(Self::MIN);
+            if LIMBS == 0 {
+                return Ok(Self::ZERO);
             }
         }
         let mut limbs = [0; LIMBS];
@@ -120,7 +342,7 @@ impl<const BITS: usize, const LIMBS: usize> TryFrom<u64> for Uint<BITS, LIMBS> {
 
 // u128 version is handled specially in because it covers two limbs.
 impl<const BITS: usize, const LIMBS: usize> TryFrom<u128> for Uint<BITS, LIMBS> {
-    type Error = ToUintError;
+    type Error = ToUintError<Self>;
 
     #[allow(clippy::cast_lossless)]
     #[allow(clippy::cast_possible_truncation)]
@@ -129,17 +351,18 @@ impl<const BITS: usize, const LIMBS: usize> TryFrom<u128> for Uint<BITS, LIMBS> 
             return Self::try_from(value as u64);
         }
         if Self::LIMBS < 2 {
-            return Err(ToUintError::ValueTooLarge(BITS));
-        }
-        let lo = value as u64;
-        let hi = (value >> 64) as u64;
-        if Self::LIMBS == 2 && hi > Self::MASK {
-            return Err(ToUintError::ValueTooLarge(BITS));
+            return Self::try_from(value as u64)
+                .and_then(|n| Err(ToUintError::ValueTooLarge(BITS, n)));
         }
         let mut limbs = [0; LIMBS];
-        limbs[0] = lo;
-        limbs[1] = hi;
-        Ok(Self::from_limbs(limbs))
+        limbs[0] = value as u64;
+        limbs[1] = (value >> 64) as u64;
+        if Self::LIMBS == 2 && limbs[1] > Self::MASK {
+            limbs[1] %= Self::MASK;
+            Err(ToUintError::ValueTooLarge(BITS, Self::from_limbs(limbs)))
+        } else {
+            Ok(Self::from_limbs(limbs))
+        }
     }
 }
 
@@ -147,7 +370,7 @@ impl<const BITS: usize, const LIMBS: usize> TryFrom<u128> for Uint<BITS, LIMBS> 
 macro_rules! impl_from_unsigned_int {
     ($uint:ty) => {
         impl<const BITS: usize, const LIMBS: usize> TryFrom<$uint> for Uint<BITS, LIMBS> {
-            type Error = ToUintError;
+            type Error = ToUintError<Self>;
 
             fn try_from(value: $uint) -> Result<Self, Self::Error> {
                 Self::try_from(value as u64)
@@ -166,11 +389,16 @@ impl_from_unsigned_int!(usize);
 macro_rules! impl_from_signed_int {
     ($int:ty, $uint:ty) => {
         impl<const BITS: usize, const LIMBS: usize> TryFrom<$int> for Uint<BITS, LIMBS> {
-            type Error = ToUintError;
+            type Error = ToUintError<Self>;
 
             fn try_from(value: $int) -> Result<Self, Self::Error> {
                 if value < 0 {
-                    Err(Self::Error::ValueNegative(BITS))
+                    Err(match Self::try_from(value as $uint) {
+                        Ok(n) | Err(ToUintError::ValueTooLarge(_, n)) => {
+                            ToUintError::ValueNegative(BITS, n)
+                        }
+                        _ => unreachable!(),
+                    })
                 } else {
                     Self::try_from(value as $uint)
                 }
@@ -187,18 +415,29 @@ impl_from_signed_int!(i128, u128);
 impl_from_signed_int!(isize, usize);
 
 impl<const BITS: usize, const LIMBS: usize> TryFrom<f64> for Uint<BITS, LIMBS> {
-    type Error = ToUintError;
+    type Error = ToUintError<Self>;
 
+    // TODO: Correctly implement wrapping.
     fn try_from(value: f64) -> Result<Self, Self::Error> {
         if value.is_nan() {
             return Err(ToUintError::NotANumber(BITS));
         }
         if value < 0.0 {
-            return Err(ToUintError::ValueNegative(BITS));
+            let wrapped = match Self::try_from(value.abs()) {
+                Ok(n) | Err(ToUintError::ValueTooLarge(_, n)) => n,
+                _ => Self::ZERO,
+            }
+            .wrapping_neg();
+            return Err(ToUintError::ValueNegative(BITS, wrapped));
         }
         #[allow(clippy::cast_precision_loss)] // BITS is small-ish
-        if value >= (Self::BITS as f64).exp2() {
-            return Err(ToUintError::ValueTooLarge(BITS));
+        let modulus = (Self::BITS as f64).exp2();
+        if value >= modulus {
+            let wrapped = match Self::try_from(value % modulus) {
+                Ok(n) | Err(ToUintError::ValueTooLarge(_, n)) => n,
+                _ => Self::ZERO,
+            };
+            return Err(ToUintError::ValueTooLarge(BITS, wrapped)); // Wrapping
         }
         if value < 0.5 {
             return Ok(Self::ZERO);
@@ -223,22 +462,28 @@ impl<const BITS: usize, const LIMBS: usize> TryFrom<f64> for Uint<BITS, LIMBS> {
         // Convert mantissa * 2^(exponent - 52) to Uint
         #[allow(clippy::cast_possible_truncation)] // exponent is small-ish
         if exponent as usize > Self::BITS + 52 {
-            return Err(ToUintError::ValueTooLarge(BITS));
+            // Wrapped value is zero because the value is extended with zero bits.
+            return Err(ToUintError::ValueTooLarge(BITS, Self::ZERO));
         }
         if exponent <= 52 {
             // Truncate mantissa
             Self::try_from(mantissa >> (52 - exponent))
         } else {
             #[allow(clippy::cast_possible_truncation)] // exponent is small-ish
-            Self::try_from(mantissa)?
-                .checked_shl(exponent as usize - 52)
-                .ok_or(ToUintError::ValueTooLarge(BITS))
+            let exponent = exponent as usize - 52;
+            let n = Self::try_from(mantissa)?;
+            let (n, overflow) = n.overflowing_shl(exponent);
+            if overflow {
+                Err(ToUintError::ValueTooLarge(BITS, n))
+            } else {
+                Ok(n)
+            }
         }
     }
 }
 
 impl<const BITS: usize, const LIMBS: usize> TryFrom<f32> for Uint<BITS, LIMBS> {
-    type Error = ToUintError;
+    type Error = ToUintError<Self>;
 
     fn try_from(value: f32) -> Result<Self, Self::Error> {
         #[allow(clippy::cast_lossless)]
@@ -253,7 +498,7 @@ impl<const BITS: usize, const LIMBS: usize> TryFrom<f32> for Uint<BITS, LIMBS> {
 macro_rules! to_value_to_ref {
     ($t:ty) => {
         impl<const BITS: usize, const LIMBS: usize> TryFrom<Uint<BITS, LIMBS>> for $t {
-            type Error = FromUintError<BITS, Self>;
+            type Error = FromUintError<Self>;
 
             fn try_from(value: Uint<BITS, LIMBS>) -> Result<Self, Self::Error> {
                 Self::try_from(&value)
@@ -265,14 +510,14 @@ macro_rules! to_value_to_ref {
 to_value_to_ref!(bool);
 
 impl<const BITS: usize, const LIMBS: usize> TryFrom<&Uint<BITS, LIMBS>> for bool {
-    type Error = FromUintError<BITS, Self>;
+    type Error = FromUintError<Self>;
 
     fn try_from(value: &Uint<BITS, LIMBS>) -> Result<Self, Self::Error> {
         if BITS == 0 {
             return Ok(false);
         }
         if value.bit_len() > 1 {
-            return Err(Self::Error::Overflow(PhantomData));
+            return Err(Self::Error::Overflow(BITS, value.bit(0), true));
         }
         Ok(value.as_limbs()[0] != 0)
     }
@@ -283,14 +528,18 @@ macro_rules! to_int {
         to_value_to_ref!($int);
 
         impl<const BITS: usize, const LIMBS: usize> TryFrom<&Uint<BITS, LIMBS>> for $int {
-            type Error = FromUintError<BITS, Self>;
+            type Error = FromUintError<Self>;
 
             fn try_from(value: &Uint<BITS, LIMBS>) -> Result<Self, Self::Error> {
                 if BITS == 0 {
                     return Ok(0);
                 }
                 if value.bit_len() > $bits {
-                    return Err(Self::Error::Overflow(PhantomData));
+                    return Err(Self::Error::Overflow(
+                        BITS,
+                        value.limbs[0] as Self,
+                        Self::MAX,
+                    ));
                 }
                 Ok(value.as_limbs()[0] as Self)
             }
@@ -310,7 +559,7 @@ to_int!(u64, 64);
 to_value_to_ref!(i128);
 
 impl<const BITS: usize, const LIMBS: usize> TryFrom<&Uint<BITS, LIMBS>> for i128 {
-    type Error = FromUintError<BITS, Self>;
+    type Error = FromUintError<Self>;
 
     #[allow(clippy::cast_lossless)] // Safe casts
     #[allow(clippy::use_self)] // More readable
@@ -318,11 +567,14 @@ impl<const BITS: usize, const LIMBS: usize> TryFrom<&Uint<BITS, LIMBS>> for i128
         if BITS == 0 {
             return Ok(0);
         }
-        if value.bit_len() > 127 {
-            return Err(Self::Error::Overflow(PhantomData));
+        if BITS <= 64 {
+            return Ok(value.as_limbs()[0] as i128);
         }
         let mut result = value.as_limbs()[0] as i128;
         result |= (value.as_limbs()[1] as i128) << 64;
+        if value.bit_len() > 127 {
+            return Err(Self::Error::Overflow(BITS, result, i128::MAX));
+        }
         Ok(result)
     }
 }
@@ -330,7 +582,7 @@ impl<const BITS: usize, const LIMBS: usize> TryFrom<&Uint<BITS, LIMBS>> for i128
 to_value_to_ref!(u128);
 
 impl<const BITS: usize, const LIMBS: usize> TryFrom<&Uint<BITS, LIMBS>> for u128 {
-    type Error = FromUintError<BITS, Self>;
+    type Error = FromUintError<Self>;
 
     #[allow(clippy::cast_lossless)] // Safe casts
     #[allow(clippy::use_self)] // More readable
@@ -338,11 +590,14 @@ impl<const BITS: usize, const LIMBS: usize> TryFrom<&Uint<BITS, LIMBS>> for u128
         if BITS == 0 {
             return Ok(0);
         }
-        if value.bit_len() > 128 {
-            return Err(Self::Error::Overflow(PhantomData));
+        if BITS <= 64 {
+            return Ok(value.as_limbs()[0] as u128);
         }
         let mut result: u128 = value.as_limbs()[0] as u128;
         result |= (value.as_limbs()[1] as u128) << 64;
+        if value.bit_len() > 128 {
+            return Err(Self::Error::Overflow(BITS, result, u128::MAX));
+        }
         Ok(result)
     }
 }
@@ -394,7 +649,7 @@ mod test {
         assert_eq!(Uint::<0, 0>::try_from(0_u64), Ok(Uint::ZERO));
         assert_eq!(
             Uint::<0, 0>::try_from(1_u64),
-            Err(ToUintError::ValueTooLarge(0))
+            Err(ToUintError::ValueTooLarge(0, Uint::ZERO))
         );
         const_for!(BITS in NON_ZERO {
             const LIMBS: usize = nlimbs(BITS);