+++ src/modular.rs
@@ -133,8 +133,9 @@ impl<const BITS: usize, const LIMBS: usize> Uint<BITS, LIMBS> {
     /// # use ruint::{uint, Uint, aliases::*};
     /// # uint!{
     /// # let modulus = 21888242871839275222246405745257275088548364400416034343698204186575808495617_U256;
-    /// let inv = (-U64::from(modulus.as_limbs()[0]).inv_ring().unwrap()).as_limbs()[0];
-    /// # assert_eq!(inv.wrapping_mul(modulus.as_limbs()[0]), u64::MAX);
+    /// let inv = U64::wrapping_from(modulus).inv_ring().unwrap().wrapping_neg().to();
+    /// let prod = 5_U256.mul_redc(6_U256, modulus, inv);
+    /// # assert_eq!(inv.wrapping_mul(modulus.wrapping_to()), u64::MAX);
     /// # assert_eq!(inv, 0xc2e1f593efffffff);
     /// # }
     /// ```
@@ -142,7 +143,6 @@ impl<const BITS: usize, const LIMBS: usize> Uint<BITS, LIMBS> {
     /// # Panics
     ///
     /// Panics if `inv` is not correct.
-    // TODO: Improve the conversion dev-ex.
     #[must_use]
     pub fn mul_redc(self, other: Self, modulus: Self, inv: u64) -> Self {
         if BITS == 0 {