+++ Changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `Pyo3` support.
 - `from` now supports `Uint` arguments.
 - `saturating_from`, `wrapping_from`, `to`, `wrapping_to`, `saturating_to`.
 - `wrapping_from_limbs_slice`, `overflowing_from_limbs_slice`, `saturating_from_limbs_slice`.