+++ Changelog.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] — 2022-10-02
+
 ### Added
 
 - Add `Pyo3` support.
@@ -146,8 +148,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- links to version -->
 
-[unreleased]: https://github.com/recmo/uint/compare/v1.3.0...HEAD
-[1.3.0]: https://github.com/recmo/uint/releases/tag/v1.2.0
+[unreleased]: https://github.com/recmo/uint/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/recmo/uint/releases/tag/v1.4.0
+[1.3.0]: https://github.com/recmo/uint/releases/tag/v1.3.0
 [1.2.0]: https://github.com/recmo/uint/releases/tag/v1.2.0
 [1.1.0]: https://github.com/recmo/uint/releases/tag/v1.1.0
 [1.0.0]: https://github.com/recmo/uint/releases/tag/v1.0.0